limit syscall to linux
