+++ cmd/net_linux.go
@@ -1,0 +1,64 @@
+//go:build (linux && !appengine)
+// +build linux,!appengine
+// Copyright (c) 2015-2021 MinIO, Inc.
+//
+// This file is part of MinIO Object Storage stack
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package cmd
+
+import (
+	"context"
+	"net"
+	"syscall"
+	"time"
+)
+
+// checkPortAvailability - check if given host and port is already in use.
+// Note: The check method tries to listen on given port and closes it.
+// It is possible to have a disconnected client in this tiny window of time.
+func checkPortAvailability(host, port, vrf string) (err error) {
+	if vrf == "" {
+		l, err := net.Listen("tcp", net.JoinHostPort(host, port))
+		if err != nil {
+			return err
+		}
+		// As we are able to listen on this network, the port is not in use.
+		// Close the listener and continue check other networks.
+		return l.Close()
+	}
+
+	// In case of VRF use: try to open in correct vrf to validate port availability.
+	lc := &net.ListenConfig{
+		Control: func(network, address string, c syscall.RawConn) error {
+			c.Control(func(fdPtr uintptr) {
+				_ = syscall.SetsockoptString(int(fdPtr), syscall.SOL_SOCKET, syscall.SO_BINDTODEVICE, vrf)
+			})
+
+			return nil
+		},
+	}
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
+	defer cancel()
+
+	l, err := lc.Listen(ctx, "tcp", net.JoinHostPort(host, port))
+	if err != nil {
+		return err
+	}
+	// As we are able to listen on this network, the port is not in use.
+	// Close the listener and continue check other networks.
+	return l.Close()
+}
